Move EphemerisCLI to its own file and improve the invalid service error
