+++ sbpy/data/ephem/__init__.py
@@ -1,0 +1,1 @@
+from .core import Ephem  # noqa: F401
