+++ sbpy/data/tests/test_ephem.py
@@ -18,7 +18,7 @@ except ImportError:  # pragma: no cover
 
 from ... import bib
 from ..orbit import Orbit
-from ..ephem import Ephem, EphemerisCLI
+from ..ephem import Ephem
 
 # retreived from Horizons on 23 Apr 2020
 CERES = {
@@ -227,113 +227,3 @@ class TestEphemFromOorb:
             oo_ephem = Ephem.from_oo(orbit, scope="basic")
             assert "sbpy.data.ephem.Ephem.from_oo" in bib.show()
         bib.reset()
-
-
-class TestEphemCLI:
-    def test_format_epochs(self):
-        """EphemerisCLI.parse_args will always have start and step.  Test
-        number, stop, and neither."""
-
-        epochs = {
-            "start": Time("2024-07-23"),
-            "step": u.Quantity(5, "day"),
-        }
-        epochs["stop"] = epochs["start"] + 10 * epochs["step"]
-
-        # test stop
-        args = SimpleNamespace(
-            start=epochs["start"], step=epochs["step"], number=None, stop=epochs["stop"]
-        )
-
-        result = EphemerisCLI._format_epochs(args)
-        for k in epochs.keys():
-            assert result[k] == epochs[k]
-
-        # test number
-        args = SimpleNamespace(
-            start=epochs["start"], step=epochs["step"], number=10, stop=None
-        )
-
-        result = EphemerisCLI._format_epochs(args)
-        for k in epochs.keys():
-            assert result[k] == epochs[k]
-
-        # no number, no stop --> number = 60
-        args = SimpleNamespace(
-            start=epochs["start"], step=epochs["step"], number=None, stop=None
-        )
-
-        result = EphemerisCLI._format_epochs(args)
-        epochs["stop"] = epochs["start"] + 60 * epochs["step"]
-        for k in epochs.keys():
-            assert result[k] == epochs[k]
-
-    def test_radec_format(self, patch_request):
-        """Test RA Dec formatting options."""
-
-        pytest.importorskip("astroquery")
-
-        cli = EphemerisCLI(
-            ["horizons", "C/1995 O1", "--start=2024-08-16", "--radec=deg"]
-        )
-
-        row = str(cli.eph.table[0]).splitlines()[-1].split()
-        assert row[2] == "339.94076"
-        assert row[3] == "-85.76646"
-
-        cli = EphemerisCLI(
-            ["horizons", "C/1995 O1", "--start=2024-08-16", "--radec=hmsdms"]
-        )
-        row = str(cli.eph.table[0]).splitlines()[-1].split()
-        assert row[2] == "22:39:45.78"
-        assert row[3] == "-85:45:59.3"
-
-    def test_start_stop_order(self, patch_request):
-        """Test start > stop."""
-
-        pytest.importorskip("astroquery")
-
-        with pytest.raises(ValueError):
-            cli = EphemerisCLI(
-                [
-                    "horizons",
-                    "C/1995 O1",
-                    "--start=2025-01-01",
-                    "--stop=2024-01-01",
-                    "--debug",
-                ]
-            )
-
-    @pytest.mark.parametrize(
-        "service, target",
-        [("horizons", "1 Ceres (A801 AA)"), ("mpc", "1"), ("miriade", "Ceres")],
-    )
-    def test_asteroid(self, service, target, patch_request):
-        """Test services with a comet designation"""
-
-        pytest.importorskip("astroquery")
-
-        cmd = f"{service} 1 --start=2024-08-16"
-        if service == "horizons":
-            cmd += " --id-type=smallbody"
-        cli = EphemerisCLI(cmd.split())
-
-        assert cli.eph["date"][0].iso == "2024-08-16 00:00:00.000"
-        assert cli.eph.meta["target"] == target
-
-    @pytest.mark.parametrize(
-        "service, target",
-        [("horizons", "2P/Encke"), ("mpc", "2P"), ("miriade", "2P")],
-    )
-    def test_comet(self, service, target, patch_request):
-        """Test services with a comet designation"""
-
-        pytest.importorskip("astroquery")
-
-        cmd = f"{service} 2P --start=2024-08-16"
-        if service == "miriade":
-            cmd += " --type=comet"
-        cli = EphemerisCLI(cmd.split())
-
-        assert cli.eph["date"][0].iso == "2024-08-16 00:00:00.000"
-        assert cli.eph.meta["target"] == target
