+++ setup.cfg
@@ -73,7 +73,7 @@ ginga.rv.plugins =
     cometaryenhancements = sbpy.ginga_plugins:setup_cometaryenhancements
 
 console_scripts =
-    sbpy-ephem = sbpy.data.ephem:EphemerisCLI.run
+    sbpy-ephem = sbpy.data.ephem.cli:EphemerisCLI.run
 
 [tool:pytest]
 minversion = 7.0
